refactor ssrf classification helpers
